\name{oKPCA}
\alias{oKPCA}
\title{Original Kernel PCA}
\description{
  Original Kernel PCA.
}
\usage{
  oKPCA(trdata, tsdata=NULL,d, kern=rbfdot(sigma=0.5))

}
\arguments{
  \item{trdata}{a n*p matrix,  the train data for KPCA.}
  \item{tsdata}{a n1*p matrix, the test data used for reconstrction error.}
  \item{d}{a positve integer, the number of principal component to be extracted.}
  \item{kern}{a kernel function generated by \link{kernlab} package, default as gaussian kernel.}
}
\details{
  This function not only compute the principal component vectors but also the reconstruction error on tsdata. If tsdata is null, the reconstruction error on trdata will be evaluated and returned.
}
\value{
  return a list with class name 'oKPCA' and including following components,
  \item{alpha}{a n*d matrix, the first d principal component vectors in each column.}
  \item{alpha_rotate}{a n*d matrix, the first d  rotated principal component vectors in each column.}
  \item{rec_err}{a numeric vector, the construction error for each sample point.}
  \item{prop_var}{a scalar, the proportion of exaplained variance.}
}
\references{
  nothing.
}
\author{
  Liu Wei
}
\note{
  nothing
}
\seealso{
  nothing
}
\examples{
  library(kernlab)
  data(iris)
  test <- sample(1:150,20)
  d <- 2
  # traditional KPCA
  kpc <- oKPCA(iris[-test,-5], d=d)
  #print the principal component vectors
  feat <- kpc$alpha_rotate
  # Sketched Kernel PCA
  fkpca <- sKPCA(iris[-test,-5], d=d, seed = 8)
  alpha_rotate <- fkpca$alpha_rotate
  par(mfrow=c(1,2))
  plot(feat,col=as.integer(iris[-test,5]),
       xlab="1st Principal Component",ylab="2nd Principal Component", main='oKPCA')
  plot(alpha_rotate[,1:2],col=as.integer(iris[-test,5]),
       xlab="1st Principal Component",ylab="2nd Principal Component", main='sKPCA')
}
\keyword{Kernel PCA}
