\name{sKPCA}
\alias{sKPCA}
\title{sketched Kernel PCA}
\description{
   Approximate Kernel PCA with Randomized Sketches.
}
\usage{
  sKPCA(trdata, tsdata=NULL,d, m=ceiling(sqrt(nrow(trdata))), Stype='ROS',  kern=rbfdot(sigma=0.5), seed=2)

}
\arguments{
  \item{trdata}{a n*p matrix,  the train data for KPCA.}
  \item{tsdata}{a n1*p matrix, the test data used for reconstrction error.}
  \item{d}{a positve integer, the number of principal component to be extracted.}
  \item{m}{the projection dimension}
  \item{Stype}{a character, the sketch type, include
    'subGaussian', 'ROS' and 'subSampling', default as ROS.}
  \item{kern}{a kernel function generated by \link{kernlab} package, default as gaussian kernel.}
  \item{seed}{a non-negative integer, specify the random seed for generating random matrix, default as 2.}
}
\details{
  This function not only compute the principal component vectors but also the reconstruction error on tsdata. If tsdata is null, the reconstruction error on trdata will be evaluated and returned.
}
\value{
  return a list with class name 'sKPCA' and including following components,
  \item{alpha}{a n*d matrix, the first d principal component vectors in each column.}
  \item{alpha_rotate}{a n*d matrix, the first d  rotated principal component vectors in each column.}
  \item{rec_err}{a numeric vector, the construction error for each sample point.}
  \item{prop_var}{a scalar, the proportion of exaplained variance.}
}
\references{
  nothing.
}
\author{
  Liu Wei
}
\note{
  nothing
}
\seealso{
  nothing
}
\examples{
  library(kernlab)
  data(iris)
  test <- sample(1:150,20)
  # traditional KPCA
  kpc <- kpca(~.,data=iris[-test,-5],kernel="rbfdot",
              kpar=list(sigma=0.2),features=2)
  #print the principal component vectors
  feat <- pcv(kpc)
  # Sketched Kernel PCA
  d <- 2
  fkpca <- sKPCA(iris[-test,-5], d=d, seed = 8)
  alpha_rotate <- fkpca$alpha_rotate
  par(mfrow=c(1,2))
  plot(feat,col=as.integer(iris[-test,5]),
       xlab="1st Principal Component",ylab="2nd Principal Component", main='KPCA')
  plot(alpha_rotate[,1:2],col=as.integer(iris[-test,5]),
       xlab="1st Principal Component",ylab="2nd Principal Component", main='sKPCA')
}
\keyword{sketch matrix}
